Lab3/requirements.txt
sqlalchemy
pandas
psycopg2-binary
scikit-learn
joblib
streamlit
pytest